MM_35115: Code clean up
